Use common to supply recipe datagen
